mini refactor
