+++ heads/src/main/java/com/dumbdogdiner/stickyapi/bukkit/item/PlayerHeadBuilder.java
@@ -2,7 +2,7 @@
  * Copyright (c) 2020-2021 DumbDogDiner <dumbdogdiner.com>. All rights reserved.
  * Licensed under the MIT license, see LICENSE for more information...
  */
-package stickyapi.bukkit.item.generator;
+package com.dumbdogdiner.stickyapi.bukkit.item;
 
 import com.dumbdogdiner.stickyapi.annotation.DoNotCall;
 import com.dumbdogdiner.stickyapi.util.http.HttpException;
