+++ heads/src/test/java/com/dumbdogdiner/stickyapi/bukkit/item/SkullBuilderTest.java
@@ -1,4 +1,4 @@
-package com.dumbdogdiner.stickyapi.bukkit.item.generator;/*
+package com.dumbdogdiner.stickyapi.bukkit.item;/*
  * Copyright (c) 2020-2021 DumbDogDiner <dumbdogdiner.com>. All rights reserved.
  * Licensed under the MIT license, see LICENSE for more information...
  */
@@ -13,7 +13,6 @@ import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
-import stickyapi.bukkit.item.generator.SkullBuilder;
 
 import org.bukkit.Material;
 import org.bukkit.inventory.ItemStack;
